+++ documentation/changelog.md
@@ -1,5 +1,9 @@
 # Change Log
 <pre>
-v2.0	Upgraded to module api version 2 for 1.8.0.1
+v3.0    Upgraded to Google Doubleclick for Publishers (version 6.2.1)
+        Added support for Interstitial ads
+        BREAKING CHANGES: See the documentation for details
+
+v1.2	Fixed crash introduced by Ti SDK 1.8.1 [MOD-405]
 
 v1.0    Initial Release